Various BugFixes
Fixed the deletion of all previous tokens when there is a multiline comment within a description delimited by quotation marks.
Removed RESERVED from keyword list as it is not necessary there.
S19 Parser fixed when there is a malformed entry that specifies more databytes than acutally exist which would previously cause an out of bounds error.