Proper input sanitation for a2l files and removed detailed parsing of if data
remove escaped characters while importing a2l file because some A2L files declare Tab-Verb Conversions like
0 "\\"
which would otherwise lead to an error in the tokenizer while trying to detect a text in quotation marks.

As it is possible to declare custom datastructres in the A2ML Section which may be used in the if data blocks the parsing for the if data structures is not checking for unexpected keywords anymore as this would lead to potential errors while the A2ML declarations aren't implemented yet.